+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/InLineContent/InLineContentOptions.cs
@@ -2,9 +2,27 @@ namespace DIPS.Mobile.UI.Components.ListItems.Options.InLineContent;
 
 public partial class InLineContentOptions : ListItemOptions
 {
-    public override void DoBind(ListItem listItem)
+    internal static void SetupDefaults(ListItem listItem)
+    {
+        listItem.ColumnDefinitions[2].Width = (GridLength)WidthProperty.DefaultValue;
+        
+        if (listItem.InLineContent is not View inLineContent)
+        {
+            return;
+        }
+        
+        inLineContent.HorizontalOptions = (LayoutOptions)HorizontalOptionsProperty.DefaultValue;
+        inLineContent.VerticalOptions = (LayoutOptions)VerticalOptionsProperty.DefaultValue;
+        
+        if ((bool)SpanOverUnderlyingContentProperty.DefaultValue)
+        {
+            Grid.SetRowSpan(inLineContent, 2);
+        }
+    }
+
+    protected override void DoBind(ListItem listItem)
     { 
-        listItem.ContainerGrid.ColumnDefinitions[2].Width = Width;
+        listItem.ColumnDefinitions[2].Width = Width;
         
         if (listItem.InLineContent is not View inLineContent)
         {